Setup and API get request
I installed all the dependecies and setup the URL for the application and the API. I also wrote the basic layout of the application, just for making the logic of the requests and prepare the project structure. I also started to write how the get request for the tools will work.