Add and Remove Tool and ReactStrap
I implemented the reactstrap lib in the application, so it reduces the verbosity of the layout code.

I also implemented the addTool and removeTool functions. I still need to treat some fields, like the link input (so it record the "http://" before the string, if it already doesn't have it) , and the tags,(so it shows the tags being added to the tags array).