Tags and some layout improvements
I implemented the tag viewer, which show the tags that are being added inside buttons. By clicking it, you remove the tag from the array, and the button list update it self. I also created the editTool button, but i haven't implemented your functionality yet. I also improved the link of each card, now, when we click the link, it opens a new tab.