Major structure changes....
While i was trying to implement the editTool function i realised that the coding was getting in a point that i could find myself in the code, so i am going to stop the adding features for now. I am going to implement a Redux state management structure, to make a pattern over the project. So, don't expect it to be functional for now, i will keep updating this.